# Robust Homotopy performance improvements