--- a/src/PowerFlows.jl
+++ b/src/PowerFlows.jl
@@ -36,6 +36,7 @@ import SparseArrays: SparseMatrixCSC, SparseVector, sparse, sparsevec, AbstractS
 import StaticArrays: MVector
 import DataStructures: OrderedDict
 import Dates
+import LineSearches: BackTracking
 
 const IS = InfrastructureSystems
 const PSY = PowerSystems