Further performance improvements: these drop the runtime of the Robust Homotopy method from 6 seconds to 2.4 seconds. The vast majority of the improvement (6s -> 2.6s) is from switching to `LineSearches.jl`'s back tracking line search.